+++ days/07/day-07.hs
@@ -1,5 +1,7 @@
 module Main where
 
+import Data.List (break)  
+
 {-
     Day 07
 
@@ -26,21 +28,59 @@ type Name = String
 type Size = Int  
 data FSItem = File Name Size | Folder Name Size [FSItem] deriving (Show)
 
--- parseTermLine parts
-parseTermLine ("$":"cd":[dir]) = "command = cd to " ++ dir
-parseTermLine ("$":["ls"])     = "command = ls"
-parseTermLine ("dir":[dir])    = "dir = " ++ dir
-parseTermLine (size:[file])    = "file = " ++ file ++ " size = " ++ size
-parseTermLine _                = error "unexpected data"
+data FSCrumb = FSCrumb Name Size [FSItem] [FSItem] deriving (Show)
+type FSZipper = (FSItem, [FSCrumb])
+
+fsUp :: FSZipper -> FSZipper  
+fsUp (item, FSCrumb name size ls rs:bs) = (Folder name size (ls ++ [item] ++ rs), bs)
+
+fsTo :: Name -> FSZipper -> FSZipper
+fsTo name (Folder folderName size items, bs) =
+    let (ls, item:rs) = break (nameIs name) items
+    in  (item, FSCrumb folderName size ls rs:bs)
+  
+nameIs :: Name -> FSItem -> Bool  
+nameIs name (Folder folderName _ _) = name == folderName  
+nameIs name (File fileName _) = name == fileName  
+
+fsNewItem :: FSItem -> FSZipper -> FSZipper
+fsNewItem item (Folder folderName size items, bs) =
+    (Folder folderName size (item:items), bs)
+
+emptyFSZipper :: FSZipper
+emptyFSZipper = (Folder "" (-1) [], [])
+
+parseTermLine :: [[Char]] -> FSZipper -> FSZipper
+parseTermLine ("$":"cd":["/"])  _        = (Folder "/" (-1) [], [])
+parseTermLine ("$":"cd":[".."]) fszipper = fsUp fszipper
+parseTermLine ("$":"cd":[dir])  fszipper = fsTo dir fszipper
+
+parseTermLine ("$":["ls"])     fszipper = fszipper
+parseTermLine ("$":_)          fszipper = error "unexpected data"
+parseTermLine ("dir":[dir])    fszipper = fsNewItem (Folder dir (-1) []) fszipper
+parseTermLine (size:[file])    fszipper = fsNewItem (File file (read size :: Int)) fszipper
+parseTermLine _                fszipper = error "unexpected data"
 
 -- parseTermHistory
--- parseTermHistory = foldl ...
-parseTermHistory termHist = map (parseTermLine . words) $ lines termHist
+parseTermHistory termHist = foldl (flip parseTermLine) emptyFSZipper
+                            $ map words $ lines termHist
+-- parseTermHistory termHist = map ((flip parseTermLine (,)) . words) $ lines termHist
 
 main :: IO ()
 main = do
   f <- readFile "input-07.txt"
 
+  -- looked to see if fs changes ... inconclusive
+  -- Data.List.sort $ filter (not . flip (elem) ["cd","$","ls","dir",".."]) $ concatMap words $ lines f
+  --    . did see dupe files/folders (e.g. "ttntt"), but are they in the same sub-dir?
+
+  -- parsing appears to be working
+  -- *Main> a = unlines $ take 15 $ lines f
+  -- *Main> a
+  -- "$ cd /\n$ ls\ndir brhvclj\ndir clnvqg\ndir dtqtvvrn\ndir lcz\ndir pcqjncwl\ndir qwvfpgl\ndir rtmj\ndir shg\ndir tcdmgwp\n$ cd brhvclj\n$ ls\n40016 mtlscfrd.gdr\ndir mvslzl\n"
+  -- *Main> parseTermHistory a
+  -- (Folder "brhvclj" (-1) [Folder "mvslzl" (-1) [],File "mtlscfrd.gdr" 40016],[FSCrumb "/" (-1) [Folder "tcdmgwp" (-1) [],Folder "shg" (-1) [],Folder "rtmj" (-1) [],Folder "qwvfpgl" (-1) [],Folder "pcqjncwl" (-1) [],Folder "lcz" (-1) [],Folder "dtqtvvrn" (-1) [],Folder "clnvqg" (-1) []] []])
+
   putStrLn $ replicate 42 '-'
-  putStrLn $ show $ Folder "dir 01" 0 [File "yes" 42]
+  putStrLn $ parseTermHistory f
   putStrLn $ replicate 42 '-'